Move graph to Statistics namespace
