+++ osu.Game.Rulesets.Tau/Statistics/PaddleDistributionGraph.cs
@@ -8,7 +8,6 @@ using osu.Framework.Graphics;
 using osu.Framework.Graphics.Colour;
 using osu.Framework.Graphics.Containers;
 using osu.Framework.Graphics.Shapes;
-using osu.Framework.Graphics.UserInterface;
 using osu.Game.Beatmaps;
 using osu.Game.Rulesets.Scoring;
 using osu.Game.Rulesets.Tau.Objects;
@@ -16,7 +15,7 @@ using osu.Game.Screens.Ranking.Expanded.Accuracy;
 using osuTK;
 using osuTK.Graphics;
 
-namespace osu.Game.Rulesets.Tau
+namespace osu.Game.Rulesets.Tau.Statistics
 {
     public class PaddleDistributionGraph : CompositeDrawable
     {
