+++ osu.Game.Rulesets.Tau/TauRuleset.cs
@@ -24,6 +24,7 @@ using osu.Game.Rulesets.Tau.Mods;
 using osu.Game.Rulesets.Tau.Replays;
 using osu.Game.Rulesets.Tau.Scoring;
 using osu.Game.Rulesets.Tau.Skinning.Legacy;
+using osu.Game.Rulesets.Tau.Statistics;
 using osu.Game.Rulesets.Tau.UI;
 using osu.Game.Rulesets.UI;
 using osu.Game.Scoring;
